requirements.txt
telethon>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
mutagen>=1.47.0
yt-dlp>=2025.12.08
pytubefix>=6.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
boto3>=1.34.0
supabase>=2.0.0

